+++ Recipe Roll.xcodeproj/xcuserdata/asduskun.xcuserdatad/xcdebugger/Breakpoints_v2.xcbkptlist
@@ -3,54 +3,4 @@
    uuid = "04C1650F-EF53-4EAE-A28C-853A2A8CE03A"
    type = "1"
    version = "2.0">
-   <Breakpoints>
-      <BreakpointProxy
-         BreakpointExtensionID = "Xcode.Breakpoint.FileBreakpoint">
-         <BreakpointContent
-            uuid = "CACDDAD6-62A3-4033-B8F0-A600F4907EF4"
-            shouldBeEnabled = "Yes"
-            ignoreCount = "0"
-            continueAfterRunningActions = "No"
-            filePath = "practice/View/Screens/AuthenticationVC.swift"
-            startingColumnNumber = "9223372036854775807"
-            endingColumnNumber = "9223372036854775807"
-            startingLineNumber = "105"
-            endingLineNumber = "105"
-            landmarkName = "didCompleteSignUp()"
-            landmarkType = "7">
-            <Locations>
-               <Location
-                  uuid = "CACDDAD6-62A3-4033-B8F0-A600F4907EF4 - 3e93c9eecae3ab81"
-                  shouldBeEnabled = "Yes"
-                  ignoreCount = "0"
-                  continueAfterRunningActions = "No"
-                  symbolName = "practice.AuthenticationVC.didCompleteSignUp() -&gt; ()"
-                  moduleName = "practice"
-                  usesParentBreakpointCondition = "Yes"
-                  urlString = "file:///Users/asduskun/Desktop/development/practice/practice/View/Screens/AuthenticationVC.swift"
-                  startingColumnNumber = "9223372036854775807"
-                  endingColumnNumber = "9223372036854775807"
-                  startingLineNumber = "106"
-                  endingLineNumber = "106"
-                  offsetFromSymbolStart = "32">
-               </Location>
-               <Location
-                  uuid = "CACDDAD6-62A3-4033-B8F0-A600F4907EF4 - 3e93c9eecae3abc7"
-                  shouldBeEnabled = "Yes"
-                  ignoreCount = "0"
-                  continueAfterRunningActions = "No"
-                  symbolName = "practice.AuthenticationVC.didCompleteSignUp() -&gt; ()"
-                  moduleName = "practice"
-                  usesParentBreakpointCondition = "Yes"
-                  urlString = "file:///Users/asduskun/Desktop/development/practice/practice/View/Screens/AuthenticationVC.swift"
-                  startingColumnNumber = "9223372036854775807"
-                  endingColumnNumber = "9223372036854775807"
-                  startingLineNumber = "108"
-                  endingLineNumber = "108"
-                  offsetFromSymbolStart = "32">
-               </Location>
-            </Locations>
-         </BreakpointContent>
-      </BreakpointProxy>
-   </Breakpoints>
 </Bucket>
