+++ practice/View/Screens/AuthenticationVC.swift
@@ -38,7 +38,7 @@ class AuthenticationVC: UIViewController {
         view.addSubview(signUpImage)
         signUpImage.frame = view.bounds
         configureUIElements()
-        presenter = AuthPresenter(authenticationVC: self, welcomeVC: nil, signUpVC: nil)
+        presenter = AuthPresenter(authenticationVC: self, welcomeVC: nil)
         
      }
